fix(api): crash on startup due to circular import
Thanks @RoboPhred for figuring out why TypeScript was happy but NodeJS was crashing with a "PrivateRouter() is not a function" error
