+++ src/controllers/api/v2/index.ts
@@ -1,3 +1,2 @@
 export * from "./actions";
 export * from "./apirouterv2";
-export * from "./publicrouter";
