--- src/controllers/api/v2/publicrouter.ts
@@ -1,5 +1,0 @@
-import { Router } from "express";
-
-const router = Router;
-
-export const PublicRouter = router;
